+++ lib/database/migrations/all_migrations.dart
@@ -23,6 +23,7 @@ import 'package:pantry_app/database/migrations/v29_inventory_unique_index.dart';
 import 'package:pantry_app/database/migrations/v2_inventories.dart';
 import 'package:pantry_app/database/migrations/v30_recipe_indexes_and_search.dart';
 import 'package:pantry_app/database/migrations/v31_serving_quantity.dart';
+import 'package:pantry_app/database/migrations/v32_scan_history.dart';
 import 'package:pantry_app/database/migrations/v3_normalize_units.dart';
 import 'package:pantry_app/database/migrations/v4_nutriscore_grade.dart';
 import 'package:pantry_app/database/migrations/v5_nutriscore_not_applicable.dart';
@@ -64,4 +65,5 @@ List<Migration> allMigrations() => [
   MigrationV29(),
   MigrationV30(),
   MigrationV31(),
+  MigrationV32(),
 ];